requirements.txt
openai
faiss-cpu
pandas
numpy
streamlit
tqdm
python-dotenv
